requirements.txt
# OPC-UA Pump Simulation Server Dependencies

# OPC-UA server implementation
asyncua>=1.0.0

# Configuration parsing
pyyaml>=6.0

# Database (SQLite with ORM)
sqlalchemy>=2.0.0

# REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Optional: for production deployment
# gunicorn>=21.0.0

# Note: The following are part of Python standard library (no install needed):
# - asyncio (async/await support)
# - logging (logging framework)
# - dataclasses (data structures)
# - pathlib (path handling)
# - json (JSON parsing)
# - math (mathematical functions)
# - random (random number generation)
# - datetime (date/time handling)
# - enum (enumerations)
# - typing (type hints)
# - sqlite3 (SQLite database - used by SQLAlchemy)
# - argparse (command line parsing)
# - signal (signal handling)
